Cleanup const generic code
* Move the code from `const_arrays.rs` to the correct `mod.rs` files
* Delete `const_arrays.rs` since they are now empty
* Delete `legacy_arrays.rs` as unused code
